# user-service/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  termsOfService: http://swagger.io/terms/
  title: User Service
  version: 0.0.1
servers:
- url: /
...
paths:
  /scalardb/join/{transactionId}:
    get:
      description: Transactional Join
      operationId: join
      parameters:
      - description: Transaction Id
        explode: false
        in: path
        name: transactionId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: successfully joined
      summary: Joins transaction
      tags:
      - scalardb
      x-accepts: application/json
      x-tags:
      - tag: scalardb
  /scalardb/prepare/{transactionId}:
    get:
      description: Transactional Prepare
      operationId: prepare
      parameters:
      - description: Transaction Id
        explode: false
        in: path
        name: transactionId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: successfully prepared
      summary: Prepares transaction
      tags:
      - scalardb
      x-accepts: application/json
      x-tags:
      - tag: scalardb
  /scalardb/rollback/{transactionId}:
    get:
      description: Transactional Rollback
      operationId: rollback
      parameters:
      - description: Transaction Id
        explode: false
        in: path
        name: transactionId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: successfully rolled back
      summary: Rollbacks transaction
      tags:
      - scalardb
      x-accepts: application/json
      x-tags:
      - tag: scalardb
  /scalardb/commit/{transactionId}:
    get:
      description: Transactional Commit
      operationId: commit
      parameters:
      - description: Transaction Id
        explode: false
        in: path
        name: transactionId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: successfully committed
      summary: Commits transaction
      tags:
      - scalardb
      x-accepts: application/json
      x-tags:
      - tag: scalardb
...